pyproject.toml
[tool.poetry]
name = "wol-redirect-proxy"
version = "1.0.1"
description = "A simple python Wake-on-LAN proxy"
authors = []
license = "MIT"
readme = "README.md"
packages = [{include = "wol_proxy", from="src"}]

[tool.poetry.scripts]
wol-proxy = "wol_proxy.app:main"


[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.94.1"
pyyaml = "^6.0"
uvicorn = "^0.21.0"
wakeonlan = "^3.0.0"
ping3 = "^4.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
